Extract history hook into custom hook
